Fixed some bugs, updated to v0.2
Fixed a bug where if the screen was turned on but not unlocked, it
would not wait to unlock before disabling the lock screen.

Fixed a bug where if the screen were locked when a trusted condition
ceased to be true, the script would re-enable trusted mode upon the
next unlock.